backend/requirements.txt
# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-Migrate>=4.0.0

# Database ORM
Flask-SQLAlchemy>=3.0.3
SQLAlchemy>=2.0.0
mysql-connector-python>=8.0.33

# Authentication & Security
Flask-JWT-Extended>=4.4.4
bcrypt>=4.0.1
python-dotenv>=1.0.0

# Face Recognition
numpy>=1.26.0
opencv-python>=4.8.0
face-recognition>=1.3.0
# dlib for face-recognition (Python 3.11 recommended)
dlib>=19.24.0
Pillow>=10.0.0

# Data Validation
marshmallow>=3.18.0
email-validator>=2.0.0

# Testing
pytest>=7.3.1
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
pydantic>=2.0.0
pandas>=2.1.0

# Production
gunicorn>=20.1.0
